pyproject.toml
[tool.poetry]
name = "osv-lib"
version = "0.0.1"
description = "A library for querying OSV (https://osv.dev) distributed vulnerability database."
authors = ["Paul Horton <paul.horton@owasp.org>"]
maintainers = ["Paul Horton <paul.horton@owasp.org>"]
packages = [
    { include = "osv" }
]
include = [
    "LICENSE"
]
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'Intended Audience :: Information Technology',
    'Intended Audience :: System Administrators',
    'Topic :: Security',
    'Topic :: Software Development',
    'Topic :: System :: Software Distribution',
    'License :: OSI Approved :: Apache Software License',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Typing :: Typed'
]
keywords = [
    "osv", "vulnerability", "vulnerability-database"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/madpah/osv-lib/issues"

[tool.poetry.dependencies]
python = "^3.6"
importlib-metadata = { version = ">= 3.4", python = "< 3.8" }
packageurl-python = "^0.9.0"
requests = "^2.20.0"
# `types-requeests` should stay in sync with `requests`
types-requests = "^2.25.1"

[tool.poetry.dev-dependencies]
tox = "^3.24.3"
autopep8 = "^1.6.0"
coverage = "^6.2"
flake8 = "^4.0.1"
isort = { version = "^5.10.0", python = ">= 3.6.1" }
mypy = ">= 0.961, < 1.00"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "pyproject.toml:version"
]
branch = "main"
upload_to_pypi = true
upload_to_release = true
build_command = "pip install poetry && poetry build"